+++ src/main/java/com/cdbd/opensource/application/messagequeue/DynamicMessageListener.java
@@ -1,5 +1,7 @@
-package com.cdbd.opensource.application;
+package com.cdbd.opensource.application.messagequeue;
 
+import com.cdbd.opensource.application.EventLogCommand;
+import com.cdbd.opensource.application.EventLogFacade;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
@@ -9,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class DynamicMessageListener {
-    private final EventLogService eventLogService;
+    private final EventLogFacade eventLogService;
 
     @Async
     public void send(String topic, Object eventLog) {
